Use `javax.annotation.processing.Generated` instead of `javax.annotation.Generated`
The former is introduced in Java9 because of split-packaging issues with `javax.annotation` package.

See https://twitter.com/rafaelcodes/status/1007225270712008704?s=20&t=ao-fnWQgJR4BMhAC312yIw
